trabalhando na exclusão de reservas
