+++ backend/src/routes/conflicts.js
@@ -1,6 +1,6 @@
 const express = require('express');
 const router = express.Router();
-const { getConflicts, acceptConflict, rejectConflict } = require('../controllers/conflictsController');
+const { getConflicts, acceptConflict, rejectConflict } = require('../controllers/conflictsController'); // Ajustado para src
 
 router.get('/', getConflicts);
 router.post('/accept', acceptConflict);
