--- src/controllers/conflictsController.js
@@ -1,9 +1,0 @@
-const express = require('express');
-const router = express.Router();
-const { getConflicts, acceptConflict, rejectConflict } = require('../controllers/conflictsController');
-
-router.get('/', getConflicts);
-router.post('/accept', acceptConflict);
-router.post('/reject', rejectConflict);
-
-module.exports = router;
