--- src/routes/massUpdate.js
@@ -1,9 +1,0 @@
-const express = require('express');
-const router = express.Router();
-const { compareSql, acceptUpdate, rejectUpdate } = require('../controllers/massUpdateController'); // Ajustado para src
-
-router.post('/compare', compareSql);
-router.post('/accept', acceptUpdate);
-router.post('/reject', rejectUpdate);
-
-module.exports = router;
